+++ kernel/kernel.c
@@ -154,14 +154,14 @@ void kmain() {
 
     create_task((void(*)(void*)) handle_mouse, "Terminal mouse handler", PRIV_KERNEL, NULL);
 
-    File* shelf_file = fs_get("system/shelf.elf");
-    if (shelf_file && !BOOT_INTO_KSHELL) {
-        task* shelf_task = exec_elf("system/shelf.elf");
-        shelf_task->privilege = PRIV_SUPER;
-    } else {
-        create_task((void(*)(void*)) shell_thread, "Shell", PRIV_KERNEL, NULL);
-    }
-    kfree((void*) shelf_file);
+    // File* shelf_file = fs_get("system/shelf.elf");
+    // if (shelf_file && !BOOT_INTO_KSHELL) {
+    //     task* shelf_task = exec_elf("system/shelf.elf");
+    //     shelf_task->privilege = PRIV_SUPER;
+    // } else {
+    //     create_task((void(*)(void*)) shell_thread, "Shell", PRIV_KERNEL, NULL);
+    // }
+    // kfree((void*) shelf_file);
 
     int s = load_sysmod("system/test_drv.sys");
     printf("Module name: %s\n", sysmods_registry[s].interface->name);