+++ makefile/lint.py
@@ -93,7 +93,23 @@ def lint():
 
             content = "".join(lines)
 
-            if "Copyright (C)" not in content:
+            license_lines = [
+                "-----------------------------------------------------------------------",
+                "Farix Operating System",
+                "Copyright (C) 2026  Faris Muhammad",
+                "This program is free software: you can redistribute it and/or modify",
+                "it under the terms of the GNU Affero General Public License as published by",
+                "the Free Software Foundation, either version 3 of the License, or",
+                "(at your option) any later version.",
+                "This program is distributed in the hope that it will be useful,",
+                "but WITHOUT ANY WARRANTY; without even the implied warranty of",
+                "MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the",
+                "GNU Affero General Public License for more details.",
+                "You should have received a copy of the GNU Affero General Public License",
+                "along with this program.  If not, see <https://www.gnu.org/licenses/>.",
+                "-----------------------------------------------------------------------",
+            ]
+            if not all(line in content for line in license_lines):
                 print(f"\x1b[91mCOPYRIGHT\x1b[0m  {file}")
 
             check_redundant_struct_tags(content, file)